requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
anywidget==0.9.18
    # via leafmap
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.4
    # via gdown
blinker==1.9.0
    # via streamlit
bqplot==0.12.45
    # via leafmap
branca==0.8.1
    # via
    #   folium
    #   ipyleaflet
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   streamlit
    #   whitebox
colour==0.1.5
    # via leafmap
comm==0.2.2
    # via ipywidgets
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via ipython
duckdb==1.3.1
    # via leafmap
executing==2.2.0
    # via stack-data
filelock==3.18.0
    # via gdown
folium==0.19.5
    # via
    #   limpeza-de-dados (pyproject.toml)
    #   leafmap
fonttools==4.58.5
    # via matplotlib
gdown==5.2.0
    # via leafmap
geographiclib==2.0
    # via geopy
geojson==3.2.0
    # via leafmap
geopy==2.4.1
    # via limpeza-de-dados (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-auth==2.40.3
    # via
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.2
    # via gspread
gspread==6.2.1
    # via
    #   limpeza-de-dados (pyproject.toml)
    #   gspread-dataframe
gspread-dataframe==4.0.0
    # via limpeza-de-dados (pyproject.toml)
idna==3.10
    # via requests
ipyevents==2.0.2
    # via leafmap
ipyfilechooser==0.6.0
    # via
    #   leafmap
    #   whiteboxgui
ipyleaflet==0.20.0
    # via leafmap
ipython==9.4.0
    # via ipywidgets
ipython-pygments-lexers==1.1.1
    # via ipython
ipytree==0.2.2
    # via whiteboxgui
ipyvue==1.11.2
    # via ipyvuetify
ipyvuetify==1.11.3
    # via leafmap
ipywidgets==8.1.7
    # via
    #   anywidget
    #   bqplot
    #   ipyevents
    #   ipyfilechooser
    #   ipyleaflet
    #   ipytree
    #   ipyvue
    #   leafmap
    #   whiteboxgui
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   altair
    #   branca
    #   folium
    #   pydeck
jsonschema==4.24.0
    # via
    #   altair
    #   pystac
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter-leaflet==0.20.0
    # via ipyleaflet
jupyterlab-widgets==3.0.15
    # via ipywidgets
kiwisolver==1.4.8
    # via matplotlib
leafmap==0.45.0
    # via limpeza-de-dados (pyproject.toml)
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via leafmap
matplotlib-inline==0.1.7
    # via ipython
narwhals==1.45.0
    # via
    #   altair
    #   plotly
numpy==2.3.1
    # via
    #   bqplot
    #   contourpy
    #   folium
    #   leafmap
    #   matplotlib
    #   pandas
    #   pydeck
    #   streamlit
oauthlib==3.3.1
    # via requests-oauthlib
packaging==24.2
    # via
    #   altair
    #   matplotlib
    #   plotly
    #   streamlit
pandas==2.2.3
    # via
    #   limpeza-de-dados (pyproject.toml)
    #   bqplot
    #   gspread-dataframe
    #   leafmap
    #   streamlit
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.3.0
    # via
    #   matplotlib
    #   streamlit
plotly==6.2.0
    # via leafmap
prompt-toolkit==3.0.51
    # via ipython
protobuf==6.31.1
    # via streamlit
psutil==7.0.0
    # via leafmap
psygnal==0.14.0
    # via anywidget
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==20.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyparsing==3.2.3
    # via matplotlib
pyshp==2.3.1
    # via leafmap
pysocks==1.7.1
    # via requests
pystac==1.13.0
    # via pystac-client
pystac-client==0.8.6
    # via leafmap
python-box==7.3.2
    # via leafmap
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
    #   pystac
    #   pystac-client
python-dotenv==1.1.1
    # via limpeza-de-dados (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via limpeza-de-dados (pyproject.toml)
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   folium
    #   gdown
    #   pystac-client
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
scooby==0.10.1
    # via leafmap
six==1.17.0
    # via
    #   gspread-dataframe
    #   python-dateutil
smmap==5.0.2
    # via gitdb
soupsieve==2.7
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
streamlit==1.45.0
    # via limpeza-de-dados (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via gdown
traitlets==5.14.3
    # via
    #   bqplot
    #   comm
    #   ipython
    #   ipywidgets
    #   matplotlib-inline
    #   traittypes
traittypes==0.2.1
    # via
    #   bqplot
    #   ipyleaflet
typing-extensions==4.14.1
    # via
    #   altair
    #   anywidget
    #   beautifulsoup4
    #   referencing
    #   streamlit
tzdata==2025.2
    # via pandas
unidecode==1.4.0
    # via limpeza-de-dados (pyproject.toml)
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
wcwidth==0.2.13
    # via prompt-toolkit
whitebox==2.3.6
    # via whiteboxgui
whiteboxgui==2.3.0
    # via leafmap
widgetsnbextension==4.0.14
    # via ipywidgets
xyzservices==2025.4.0
    # via
    #   folium
    #   ipyleaflet
    #   leafmap
pre-commit-4.3.0